# docs/swagger.yaml
basePath: /
definitions:
  handler.CreateSubscriptionRequest:
    properties:
      end_date:
        example: 12-2025
        type: string
      price:
        example: 400
        minimum: 0
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  handler.ErrorResponse:
    properties:
      error:
        example: invalid request
        type: string
    type: object
  handler.SubscriptionResponse:
    properties:
      created_at:
        type: string
      end_date:
        example: "2025-12-01"
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      price:
        example: 400
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: "2025-07-01"
        type: string
      updated_at:
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  handler.TotalCostResponse:
    properties:
      total:
        example: 12000
        type: integer
    type: object
  handler.UpdateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        minimum: 0
        type: integer
      service_name:
        type: string
      start_date:
        type: string
    required:
    - price
    - service_name
    - start_date
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for managing online subscriptions
  title: Subscriptions API
  version: "1.0"
paths:
  /api/v1/subscriptions:
    get:
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      - description: Filter by service name
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.SubscriptionResponse'
            type: array
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      parameters:
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/handler.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create subscription
      tags:
      - subscriptions
  /api/v1/subscriptions/{id}:
    delete:
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SubscriptionResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update subscription
      tags:
      - subscriptions
  /api/v1/subscriptions/total-cost:
    get:
      parameters:
      - description: Start period YYYY-MM-DD
        in: query
        name: start_period
        required: true
        type: string
      - description: End period YYYY-MM-DD
        in: query
        name: end_period
        required: true
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      - description: Filter by service name
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TotalCostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Calculate total cost
      tags:
      - subscriptions
swagger: "2.0"